[Serialization] Widen extension-table dataLength to uint32. rdar://181101562

+++ lib/Serialization/ModuleFileCoreTableInfo.h
@@ -113,7 +113,7 @@ public:
 
   static std::pair<unsigned, unsigned> ReadKeyDataLength(const uint8_t *&data) {
     unsigned keyLength = readNext<uint16_t>(data);
-    unsigned dataLength = readNext<uint16_t>(data);
+    unsigned dataLength = readNext<uint32_t>(data);
     return { keyLength, dataLength };
   }
